(feature) extract create post buttons; set space in PostDialog
This is a first commit for creating posts with the "Create Post" buttons from side and bottom navigation.

As there are multiple placed to trigger the dialog from, I decided to move the logic for figuring the current space into the dialog itself.

Next step would be to handle the case when there is no space selected and allow the user to select the space rather than pre-defining one.